fix sheet_test.dart for cupertino ; move material test to subfolder for now
